Rename 'new' -> '' proof fixtures in JS
