--- ethereum/oracles/src/script/noir_fixtures/new_receipt_proof.ts
@@ -1,7 +1,0 @@
-import { Proof } from '../../ethereum/proof.js';
-import { receiptProofConfigM } from '../../noir/oracles/rpc/common/proofConfig/receipt.js';
-import { createNewTopLevelProofFixture } from './new_proof.js';
-
-export function createNewReceiptProofFixture(proof: Proof): string {
-  return createNewTopLevelProofFixture(proof, receiptProofConfigM);
-}
