+++ ethereum/oracles/src/script/noir_fixtures/transaction_proof.ts
@@ -1,15 +1,7 @@
 import { Proof } from '../../ethereum/proof.js';
-import { LEGACY_MAX_TX_ENCODED_LEN, txProofConfigM } from '../../noir/oracles/rpc/common/proofConfig/tx.js';
-import { BYTE_HEX_LEN } from '../../util/const.js';
-import { createProofFixture } from './proof.js';
+import { txProofConfigM } from '../../noir/oracles/common/proofConfig/tx.js';
+import { createTopLevelProofFixture } from './proof.js';
 
 export function createTransactionProofFixture(proof: Proof): string {
-  return createProofFixture(
-    proof,
-    txProofConfigM.maxKeyLen * BYTE_HEX_LEN,
-    LEGACY_MAX_TX_ENCODED_LEN,
-    txProofConfigM.maxProofLen,
-    'dep::proof::trie_proof::TrieProof',
-    'TrieProof'
-  );
+  return createTopLevelProofFixture(proof, txProofConfigM);
 }
