+++ ethereum/oracles/src/script/prepareMerkleProofFixtures.ts
@@ -4,8 +4,8 @@ import { PROOF_FIXTURES } from '../fixtures/merkleProofsConfig.js';
 import { assert, encodeHexStringToArray } from '../main.js';
 import { hasDuplicates } from '../util/array.js';
 import { bytesToHex } from 'viem';
-import { createNewTopLevelProofFixtureWithRoot } from './noir_fixtures/new_proof.js';
-import { getProofConfig } from '../noir/oracles/rpc/common/proofConfig.js';
+import { createTopLevelProofFixtureWithRoot } from './noir_fixtures/proof.js';
+import { getProofConfig } from '../noir/oracles/common/proofConfig.js';
 import { BYTE_HEX_LEN } from '../util/const.js';
 import { encodeUint8Array } from '../noir/oracles/common/encode.js';
 
@@ -45,7 +45,7 @@ for (const fixtureName in PROOF_FIXTURES) {
   const config = getProofConfig(maxKeyLen, MAX_VALUE_LEN, MAX_DEPTH);
   await writeFile(
     `${NOIR_PROOF_FIXTURES_DIRECTORY}/${fixtureName}.nr`,
-    createNewTopLevelProofFixtureWithRoot(proofFixture, root, config)
+    createTopLevelProofFixtureWithRoot(proofFixture, root, config)
   );
 
   fixtureModule += `mod ${fixtureName};\n`;
