+++ ethereum/oracles/src/script/prepareNoirHistoryAPIFixtures.ts
@@ -12,10 +12,10 @@ import { createTransactionFixture } from './noir_fixtures/transaction.js';
 import { getTxProof } from '../ethereum/txProof.js';
 import { createReceiptFixture } from './noir_fixtures/receipt.js';
 import { createLogFixture } from './noir_fixtures/log.js';
-import { createNewReceiptProofFixture } from './noir_fixtures/new_receipt_proof.js';
-import { createNewTransactionProofFixture } from './noir_fixtures/new_transaction_proof.js';
-import { createNewStateProofFixture } from './noir_fixtures/new_state_proof.js';
-import { createNewStorageProofFixture } from './noir_fixtures/new_storage_proof.js';
+import { createReceiptProofFixture } from './noir_fixtures/receipt_proof.js';
+import { createTransactionProofFixture } from './noir_fixtures/transaction_proof.js';
+import { createStateProofFixture } from './noir_fixtures/state_proof.js';
+import { createStorageProofFixture } from './noir_fixtures/storage_proof.js';
 import { createStorageFixture } from './noir_fixtures/storage.js';
 
 const INDEX_NOT_FOUND = -1;
@@ -59,15 +59,12 @@ for (const chain in HISTORY_API_FIXTURES) {
         });
 
         await writeFile(join(modulePath, 'account.nr'), createAccountFixture(stateProof));
-        await writeFile(join(modulePath, 'state_proof_new.nr'), createNewStateProofFixture(stateProof));
-        fixtureModules.push('account', 'state_proof_new');
+        await writeFile(join(modulePath, 'state_proof.nr'), createStateProofFixture(stateProof));
+        fixtureModules.push('account', 'state_proof');
         if (storageKeys) {
           await writeFile(join(modulePath, 'storage.nr'), createStorageFixture(stateProof));
-          await writeFile(
-            join(modulePath, 'storage_proof_new.nr'),
-            createNewStorageProofFixture(stateProof.storageProof)
-          );
-          fixtureModules.push('storage', 'storage_proof_new');
+          await writeFile(join(modulePath, 'storage_proof.nr'), createStorageProofFixture(stateProof.storageProof));
+          fixtureModules.push('storage', 'storage_proof');
         }
       }
 
@@ -78,8 +75,8 @@ for (const chain in HISTORY_API_FIXTURES) {
         const receipt = blockReceipts[txIdx];
         const txReceiptProof = await getReceiptProof(block, blockReceipts, txIdx);
         await writeFile(join(modulePath, 'receipt.nr'), createReceiptFixture(receipt));
-        await writeFile(join(modulePath, 'receipt_proof_new.nr'), createNewReceiptProofFixture(txReceiptProof));
-        fixtureModules.push('receipt_proof_new', 'receipt');
+        await writeFile(join(modulePath, 'receipt_proof.nr'), createReceiptProofFixture(txReceiptProof));
+        fixtureModules.push('receipt_proof', 'receipt');
 
         const tx = await client.getTransaction({ hash: transactionHash });
         await writeFile(join(modulePath, 'transaction.nr'), createTransactionFixture(tx));
@@ -90,8 +87,8 @@ for (const chain in HISTORY_API_FIXTURES) {
           blockWithTransactions.transactionsRoot,
           txIdx
         );
-        await writeFile(join(modulePath, 'transaction_proof_new.nr'), createNewTransactionProofFixture(txProof));
-        fixtureModules.push('transaction_proof_new');
+        await writeFile(join(modulePath, 'transaction_proof.nr'), createTransactionProofFixture(txProof));
+        fixtureModules.push('transaction_proof');
         if (logIdx !== undefined) {
           const log = receipt.logs[logIdx];
           await writeFile(join(modulePath, 'log.nr'), createLogFixture(log, logIdx));
